Fix energy demo
